RESTEASY-1636 part2
