+++ security-legacy/login-module-authenticator/src/main/java/org/jboss/security/web/ThreadContext.java
