+++ security-legacy/pom.xml
@@ -25,6 +25,7 @@
 				<module>keystone/keystone-core</module>
 				<module>keystone/keystone-as7</module>
 				<module>keystone/keystone-as7-modules</module>
+                <module>login-module-authenticator</module>			
             </modules>
         </profile>
         <profile>
@@ -37,6 +38,7 @@
 			    <module>keystone/keystone-core</module>
 			    <module>keystone/keystone-as7</module>
 			    <module>keystone/keystone-as7-modules</module>
+                <module>login-module-authenticator</module>				
 			    <!--
 			    <module>skeleton-keystone-idm/example/skeleton-key-server</module>
 			    <module>skeleton-keystone-idm/example/some-app</module>
